Added car buy button interactable state controller;

+++ Assets/Scripts/ScriptableObjects/Scripts/CarSpecification.cs
@@ -9,12 +9,12 @@ public class CarSpecification : ScriptableObject
     [SerializeField] private float _engineTorque;
     [SerializeField] private float _breakTorque;
     [SerializeField] private int _gears;
-    [SerializeField] private float _price;
+    [SerializeField] private int _price;
     
     public Pools CarPool => _carPool;
     public float MaxSpeed => _maxSpeed;
     public float EngineTorque => _engineTorque;
     public float BreakTorque => _breakTorque;
     public int Gears => _gears;
-    public float Price => _price;
+    public int Price => _price;
 }
